refactor(services): split score_cache and dashboard into focused modules
score_cache.py becomes a facade over _score_cache_db (connection,
schema, epoch purge), _score_cache_store (row/summary read-write),
_score_cache_fetch (read-through caches) and _score_cache_epoch.
dashboard.py delegates to _dashboard_cache (run-dim LRU),
_dashboard_history and _dashboard_response.

The single-flight machinery from #1077 moves with its code: the open
lock into _score_cache_db, the per-key in-flight registry into
_score_cache_fetch. The accumulated-persist spy now patches
_score_cache_fetch, where the write is resolved.

No behavior change.
