+++ src/quodeq/services/dashboard.py
@@ -1,166 +1,61 @@
-"""Dashboard and accumulated-view logic, split from action_provider_fs."""
+"""Dashboard and accumulated-view logic, split from action_provider_fs.
+
+This module owns the *selected run*: resolving which run the request means,
+rescoring its dimensions against project-wide suppressions, and driving the
+three collaborators that do the rest.
+
+- ``_dashboard_cache``    — run-dimension LRU config, shared cache, fetchers
+- ``_dashboard_history``  — previous scores, stale dimensions, trend series
+- ``_dashboard_response`` — camelCase serialization of the response
+
+Their symbols are re-exported below so existing import paths keep resolving.
+"""
 from __future__ import annotations
 
-import logging
-import os
 import threading
 from collections import OrderedDict
-from dataclasses import dataclass, replace
 from pathlib import Path
 from typing import Any, Callable
 
-from quodeq.core.scoring.params import DEFAULT_PARAMS, ScoringParams
-from quodeq.core.types import DimensionResult, DimensionSummary, to_camel_dict
+from quodeq.core.scoring.params import ScoringParams
+from quodeq.core.types import DimensionResult
 
-from quodeq.data.fs.report_parser.grades import calculate_trend, summarize_dimensions
+from quodeq.data.fs.report_parser.grades import summarize_dimensions
 from quodeq.data.fs.report_parser.runs import RunInfo, list_runs, read_run_data
-from quodeq.services._cache import make_lru_dimension_fetcher
-from quodeq.services._dashboard_stale import collect_stale_dimensions
-from quodeq.services._dashboard_trend import build_accumulated_trend
-from quodeq.services._trend_fetcher import make_trend_fetcher
-from quodeq.services.scoring_view import is_eligible_for_default_view, select_trend_runs
+from quodeq.services.scoring_view import is_eligible_for_default_view
 from quodeq.services.dismissed import filter_dismissed_from_dimensions
 from quodeq.shared.validation import validate_path_segment
 from quodeq.data.fs.suppression_rules import load_suppression_rules
 
-_logger = logging.getLogger(__name__)
+from quodeq.services._dashboard_cache import (  # noqa: F401
+    DashboardCacheConfig,
+    _DEFAULT_RUN_DIM_CACHE_MAX,
+    _make_run_dimension_fetcher,
+    _run_dim_cache_max,
+    _SHARED_RUN_DIM_CACHE,
+    _SHARED_RUN_DIM_LOCK,
+    clear_shared_dimension_cache,
+    create_dimension_cache,
+)
+from quodeq.services._dashboard_history import (  # noqa: F401
+    _DashboardPayload,
+    _DEFAULT_MAX_HISTORY_RUNS,
+    _SKIP_GRADES,
+    _SelectedRunContext,
+    _collect_previous_scores,
+    _compute_dashboard_payload,
+    _enrich_dimensions_with_trend,
+    _max_history_runs,
+)
+from quodeq.services._dashboard_response import (  # noqa: F401
+    _attach_dismissed_count_to_dim,
+    _attach_exit_reason_to_dim,
+    _build_dashboard_result,
+    _read_run_exit_reason,
+    _slim_history_dim,
+)
 
-
-@dataclass
-class DashboardCacheConfig:
-    """Optional cache overrides for build_dashboard (mirrors AccumulatedCacheConfig)."""
-    cache: OrderedDict[tuple, list[DimensionResult]] | None = None
-    lock: threading.Lock | None = None
-    max_size: int | None = None
-
-
-_SKIP_GRADES = {"NA", "N/A", "INSUFFICIENT"}
-
-# Maximum number of historical runs scanned for trend, previous scores, and
-# stale dimensions. The full run list is still returned in availableRuns (metadata
-# only, no disk reads) so users can navigate to older runs directly.
 _LATEST_RUN = "latest"
-_DEFAULT_MAX_HISTORY_RUNS = 100
-
-
-def _max_history_runs(env: dict[str, str] | None = None) -> int:
-    """Return the history-scan ceiling, honouring QUODEQ_MAX_HISTORY_RUNS."""
-    raw = (env or os.environ).get("QUODEQ_MAX_HISTORY_RUNS")
-    if not raw:
-        return _DEFAULT_MAX_HISTORY_RUNS
-    try:
-        value = int(raw)
-    except ValueError:
-        return _DEFAULT_MAX_HISTORY_RUNS
-    return value if value > 0 else _DEFAULT_MAX_HISTORY_RUNS
-
-
-_DEFAULT_RUN_DIM_CACHE_MAX = 256
-
-
-def _run_dim_cache_max(override: int | None = None, env: dict[str, str] | None = None) -> int:
-    """Return the run-dimension cache size limit. *override* bypasses env for testing."""
-    if override is not None:
-        return override
-    try:
-        return int((env or os.environ).get("QUODEQ_RUN_DIM_CACHE_MAX", str(_DEFAULT_RUN_DIM_CACHE_MAX)))
-    except (ValueError, TypeError):
-        return _DEFAULT_RUN_DIM_CACHE_MAX
-
-
-# Module-level shared cache for run-dimension data. Without this, every
-# dashboard request used a fresh cache (created in _make_run_dimension_fetcher
-# below), so re-fetching the same project's history (which collect_stale_dimensions
-# / _collect_previous_scores / build_accumulated_trend all walk) cost ~750ms
-# per request even on warm calls. The shared cache eliminates the cross-request
-# I/O without compromising the per-request consistency guarantees (the cache
-# is keyed by (reports_root, project, run_id, suppression_version) so a
-# dismiss/delete produces a new key and never serves a pre-suppression score,
-# and runs are immutable once finalized).
-#
-# Tests that need isolation can pass an explicit DashboardCacheConfig.
-_SHARED_RUN_DIM_CACHE, _SHARED_RUN_DIM_LOCK = OrderedDict(), threading.Lock()
-
-
-def create_dimension_cache() -> tuple[OrderedDict[tuple, list[DimensionResult]], threading.Lock]:
-    """Create the default run-dimension LRU cache and its lock.
-
-    Override this factory to plug in a shared backend (e.g. a Redis-backed
-    OrderedDict wrapper) for multi-worker deployments.  The returned
-    ordered-dict must support ``move_to_end``, ``popitem(last=False)``,
-    and standard ``__getitem__``/``__setitem__``/``__contains__``.
-    """
-    return OrderedDict(), threading.Lock()
-
-
-def clear_shared_dimension_cache() -> None:
-    """Drop all cached run-dimension data (e.g. after a formula change)."""
-    with _SHARED_RUN_DIM_LOCK:
-        _SHARED_RUN_DIM_CACHE.clear()
-
-
-def _collect_previous_scores(
-    runs: list[RunInfo], selected_index: int, selected_dim_names: set[str],
-    get_run_dimensions: Callable[[str], list[DimensionResult]],
-) -> dict[str, DimensionResult]:
-    """Find the most recent previous score for each dimension in the selected run."""
-    previous_by_dimension: dict[str, DimensionResult] = {}
-    for older_idx in range(selected_index + 1, len(runs)):
-        run_dimensions = get_run_dimensions(runs[older_idx].run_id)
-        for dim in run_dimensions:
-            dim_name = dim.dimension
-            if not dim_name or dim_name not in selected_dim_names:
-                continue
-            grade = dim.overall_grade
-            if not grade or str(grade).upper() in _SKIP_GRADES:
-                continue
-            if dim_name not in previous_by_dimension:
-                previous_by_dimension[dim_name] = replace(dim, run_id=runs[older_idx].run_id)
-    return previous_by_dimension
-
-
-def _enrich_dimensions_with_trend(
-    selected_dimensions: list[DimensionResult], previous_by_dimension: dict[str, DimensionResult]
-) -> list[DimensionResult]:
-    """Attach trend and previous-run data to each selected dimension."""
-    result: list[DimensionResult] = []
-    for dim in selected_dimensions:
-        previous = previous_by_dimension.get(dim.dimension or "")
-        trend = calculate_trend(dim.overall_score, previous.overall_score if previous else None)
-        result.append(
-            replace(
-                dim,
-                trend=trend,
-                previous_run_id=previous.run_id if previous else None,
-                previous_score=previous.overall_score if previous else None,
-            )
-        )
-    return result
-
-
-def _make_run_dimension_fetcher(
-    reports_root: Path,
-    project: str,
-    cache: OrderedDict[tuple, list[DimensionResult]] | None = None,
-    lock: threading.Lock | None = None,
-    max_size: int | None = None,
-    version: str = "",
-) -> Callable[[str], list[DimensionResult]]:
-    """Return a cached fetcher for run dimension data (LRU, bounded).
-
-    Defaults to the module-level shared cache so reads of the same run's
-    dimensions across requests reuse work. *version* scopes the cache key to the
-    project's suppression state so a dismiss/delete invalidates it. Tests pass
-    explicit cache/lock to isolate state.
-    """
-    return make_lru_dimension_fetcher(
-        reports_root,
-        project,
-        cache if cache is not None else _SHARED_RUN_DIM_CACHE,
-        lock if lock is not None else _SHARED_RUN_DIM_LOCK,
-        max_size if max_size is not None else _run_dim_cache_max(),
-        version=version,
-    )
 
 
 def _rescore_run_dimensions(
@@ -231,146 +126,6 @@ def _make_status_aware_fetcher(
     return fetch
 
 
-@dataclass
-class _DashboardPayload:
-    """Pre-computed parts for the dashboard response."""
-    selected_summary: DimensionSummary
-    trend: list[dict[str, Any]]
-    dimensions_with_trend: list[DimensionResult]
-    previous_by_dimension: dict[str, DimensionResult]
-    stale_previous_by_dimension: dict[str, DimensionResult]
-    stale_dimensions: list[DimensionResult]
-
-
-def _read_run_exit_reason(reports_root: Path, project: str, run_id: str) -> str | None:
-    """Return the run's ``status.json`` ``exit_reason``, or ``None`` if absent.
-
-    Used by the dashboard to surface deadline-truncated runs to the UI:
-    the "Partial" badge on each DimensionGaugeCard fires when the run
-    didn't complete naturally (e.g. ``exit_reason="deadline"`` from a
-    timeout, or ``"failure_streak"`` from repeated failures).
-    """
-    import json as _json  # noqa: PLC0415
-    status_path = reports_root / project / run_id / "status.json"
-    if not status_path.is_file():
-        return None
-    try:
-        with status_path.open("r", encoding="utf-8") as fp:
-            data = _json.load(fp)
-    except (OSError, ValueError):
-        return None
-    reason = data.get("exit_reason")
-    return reason if isinstance(reason, str) else None
-
-
-def _attach_exit_reason_to_dim(
-    dim_dict: dict[str, Any], run_exit_reason: str | None,
-) -> dict[str, Any]:
-    """Add ``exitReason`` to a serialized dimension dict.
-
-    Preference order: per-dim exit_reason (if present on the dim) wins over
-    the run-level value. Either way, the chosen reason is exposed to the UI
-    as ``exitReason``. Falls back to no key when both are absent (legacy).
-    """
-    per_dim = dim_dict.get("exit_reason") or dim_dict.get("exitReason")
-    chosen = per_dim or run_exit_reason
-    if chosen is None:
-        # Drop the snake_case key if present, to keep the response clean.
-        if "exit_reason" in dim_dict:
-            out = dict(dim_dict)
-            out.pop("exit_reason", None)
-            return out
-        return dim_dict
-    out = dict(dim_dict)
-    out.pop("exit_reason", None)
-    out["exitReason"] = chosen
-    return out
-
-
-def _attach_dismissed_count_to_dim(
-    dim_dict: dict[str, Any], dismissed_counts: dict[str, int],
-    suppressed_counts: dict[str, int] | None = None,
-) -> dict[str, Any]:
-    """Add ``dismissedCount`` / ``suppressedCount`` to a dimension dict when > 0.
-
-    Both explain the gap between "what the scan found" and "what the view
-    shows". ``dismissedCount`` covers the dismissed filter alone;
-    ``suppressedCount`` covers dismissals *and* deletions, so it is the total
-    the UI reports. They differ sharply on projects with a triage history:
-    deletions suppress a whole principle across a file and accumulate over
-    many runs, so a scan can re-find several times what the report displays.
-    Omitted when nothing was filtered, mirroring the exitReason convention.
-    """
-    key = dim_dict.get("dimension") or ""
-    out = dim_dict
-    count = dismissed_counts.get(key, 0)
-    if count > 0:
-        out = {**out, "dismissedCount": count}
-    total = (suppressed_counts or {}).get(key, 0)
-    if total > 0:
-        out = {**out, "suppressedCount": total}
-    return out
-
-
-def _slim_history_dim(dim: DimensionResult) -> dict[str, Any]:
-    """Serialize a history-context dimension without its finding bodies.
-
-    The previousByDimension / stalePreviousByDimension / staleDimensions keys
-    exist to carry scores, grades, and provenance (run id, dates) for trend
-    context; the UI reads only the scalar fields inlined on each selected-run
-    dimension (previousScore, trend, stale, fromRunId). No consumer reads the
-    violations/compliance arrays from these keys, yet on large projects they
-    dominated the payload: for a 201-run project, an old run's dashboard was
-    19.9 MB of which these three keys carried 18.6 MB of finding bodies.
-    Totals keep the counts; only the bodies are dropped.
-    """
-    return to_camel_dict(replace(dim, violations=[], compliance=[]))
-
-
-def _build_dashboard_result(
-    project: str,
-    runs: list[RunInfo],
-    selected_run: RunInfo,
-    payload: _DashboardPayload,
-    *,
-    exit_reason: str | None = None,
-    dismissed_counts: dict[str, int] | None = None,
-    suppressed_counts: dict[str, int] | None = None,
-) -> dict[str, Any]:
-    """Assemble the final dashboard response dict from pre-computed parts."""
-    dim_dicts = [
-        _attach_dismissed_count_to_dim(
-            _attach_exit_reason_to_dim(to_camel_dict(d), exit_reason),
-            dismissed_counts or {},
-            suppressed_counts or {},
-        )
-        for d in payload.dimensions_with_trend
-    ]
-    return {
-        "project": project,
-        "availableRuns": [
-            {"runId": item.run_id, "dateISO": item.date_iso, "dateLabel": item.date_label, "status": item.status}
-            for item in runs
-        ],
-        "selectedRun": {
-            "runId": selected_run.run_id,
-            "dateISO": selected_run.date_iso,
-            "dateLabel": selected_run.date_label,
-            "exitReason": exit_reason,
-        },
-        "summary": {
-            **to_camel_dict(payload.selected_summary),
-            "dateISO": selected_run.date_iso,
-            "dateLabel": selected_run.date_label,
-        },
-        "trend": payload.trend,
-        "dimensions": dim_dicts,
-        "previousByDimension": {k: _slim_history_dim(v) for k, v in payload.previous_by_dimension.items()},
-        "stalePreviousByDimension": {k: _slim_history_dim(v) for k, v in payload.stale_previous_by_dimension.items()},
-        "staleDimensions": [_slim_history_dim(d) for d in payload.stale_dimensions],
-    }
-
-
 # Fallback order for the "latest" default run when none is complete. Each
 # tier is tried newest-first; a failed run is only headlined when nothing
 # else remains (handled after this list). Complete mirrors the Overview's
@@ -423,94 +178,6 @@ def _resolve_selected_run(runs: list[RunInfo], run: str) -> tuple[RunInfo, int]:
     if selected_index is None:
         raise RuntimeError(f"Run {selected_run.run_id!r} disappeared from the run list unexpectedly.")
     return selected_run, selected_index
-
-
-@dataclass(frozen=True)
-class _SelectedRunContext:
-    """Pre-resolved data for the selected run in a dashboard request."""
-    run: RunInfo
-    index: int
-    dimensions: list[DimensionResult]
-    summary: DimensionSummary
-
-
-def _compute_dashboard_payload(
-    reports_root: Path, project: str, runs: list[RunInfo],
-    ctx: _SelectedRunContext, cc: DashboardCacheConfig,
-    params: ScoringParams = DEFAULT_PARAMS,
-) -> _DashboardPayload:
-    """Compute history-dependent parts of the dashboard response."""
-    selected_dim_names = {d.dimension for d in ctx.dimensions}
-    # Shared trend rule (scoring_view.select_trend_runs): cancelled/failed
-    # runs are excluded — misleading history points. They remain visible in
-    # availableRuns for the UI.
-    scoreable_runs = select_trend_runs(runs)
-    # Re-find the selected run's index inside scoreable_runs. ctx.index is
-    # the index in the full unfiltered run list, which can exceed
-    # len(history_runs) when cancelled/failed runs sit above the selected
-    # run. Passing the wrong index to collect_stale_dimensions /
-    # _collect_previous_scores caused IndexError on history_runs[newer_idx].
-    selected_in_scoreable = next(
-        (i for i, r in enumerate(scoreable_runs) if r.run_id == ctx.run.run_id),
-        None,
-    )
-    max_history = _max_history_runs()
-    if selected_in_scoreable is None:
-        # Selected run was cancelled/failed (so it's not in scoreable_runs).
-        # Treat the entire scoreable history as "older" runs relative to it.
-        history_runs = scoreable_runs[:max_history]
-        history_index = len(history_runs)
-    else:
-        history_runs = scoreable_runs[:max(max_history, selected_in_scoreable + 1)]
-        history_index = selected_in_scoreable
-    # History fetcher: cache-backed, dismiss-adjusted, SCALAR-only -- the same
-    # fetcher the /scores endpoint uses. The three consumers below
-    # (_collect_previous_scores, collect_stale_dimensions, build_accumulated_trend)
-    # read only per-run scalars (dimension + overallScore + overallGrade), not
-    # the full violations. Reading + rescoring FULL data for every history run
-    # (up to _max_history_runs()) was the ~2s cost this replaces.
-    #
-    # In-progress freshness is preserved: the fast path re-reads each request
-    # (fresh per-call cache), and the heavy path's cacheable_run_ids guard makes
-    # in-progress runs compute-through without persisting a partial set. Stale-
-    # partial detection is preserved inside read_run_scalars, which falls back to
-    # full read_run_data whenever the SQL scalar projection disagrees with the
-    # on-disk evaluation/*.json count -- the same self-heal the old status-aware
-    # fetcher did via _count_eval_files. The dismiss-adjustment (Bug B) stays;
-    # it is now cached rather than recomputed on every request.
-    cacheable_run_ids = {r.run_id for r in history_runs if r.status == "complete"}
-    # Key the in-memory dimension cache by the project's suppression state so a
-    # dismiss/delete (or a formula change) invalidates warmed entries and no
-    # read path can serve a pre-dismiss score. ``score_cache_version`` already
-    # hashes dismissed + deleted keys + params. This fetcher is SHARED across the
-    # whole history window (previous-scores, stale-dimensions, and the trend all
-    # iterate many runs through it), so we keep the global project-scoped version
-    # here rather than a per-run scoped one -- per-run scoping only makes sense
-    # when a single run is in play, which this path is not.
-    from quodeq.services.score_cache import score_cache_version  # noqa: PLC0415
-    dim_cache_version = score_cache_version(reports_root / project, params)
-    get_run_dimensions = make_trend_fetcher(
-        reports_root, project, params=params, cacheable_run_ids=cacheable_run_ids,
-        max_history=max_history,
-        base_fetcher_factory=lambda rr, proj: _make_run_dimension_fetcher(
-            rr, proj, cache=cc.cache, lock=cc.lock, max_size=cc.max_size,
-            version=dim_cache_version,
-        ),
-    )
-    previous_by_dimension = _collect_previous_scores(
-        history_runs, history_index, selected_dim_names, get_run_dimensions,
-    )
-    stale_dimensions, stale_previous_by_dimension = collect_stale_dimensions(
-        history_runs, history_index, selected_dim_names, get_run_dimensions,
-    )
-    return _DashboardPayload(
-        selected_summary=ctx.summary,
-        trend=build_accumulated_trend(history_runs, get_run_dimensions, params=params),
-        dimensions_with_trend=_enrich_dimensions_with_trend(ctx.dimensions, previous_by_dimension),
-        previous_by_dimension=previous_by_dimension,
-        stale_previous_by_dimension=stale_previous_by_dimension,
-        stale_dimensions=stale_dimensions,
-    )
 
 
 def build_dashboard(
@@ -586,3 +253,11 @@ def build_dashboard(
         exit_reason=exit_reason, dismissed_counts=dismissed_counts,
         suppressed_counts=suppressed_counts,
     )
+
+
+__all__ = [
+    "DashboardCacheConfig",
+    "build_dashboard",
+    "clear_shared_dimension_cache",
+    "create_dimension_cache",
+]
