+++ src/quodeq/services/score_cache.py
@@ -3,253 +3,49 @@
 Keyed by a content-hash version of the project's dismissals/deletions + grade
 params, so any change auto-invalidates. Disposable/best-effort: a corrupt or
 older-schema db is rebuilt, and any cache error falls through to recompute.
+
+This module owns the version hashes (the cache's invalidation contract) and is
+the facade every caller imports. The mechanics live in three submodules:
+``_score_cache_db`` (connection/schema), ``_score_cache_store`` (per-table
+reads/writes) and ``_score_cache_fetch`` (read-through wrappers).
 """
 from __future__ import annotations
 
 import hashlib
 import json
-import logging
 import sqlite3
-import threading
-from contextlib import contextmanager
-from contextvars import ContextVar
 from pathlib import Path
-from typing import Callable, Iterator
 
 from quodeq.core.scoring.params import ScoringParams
-from quodeq.core.types import DimensionResult
 from quodeq.services.deleted import deleted_keys
 from quodeq.services.dismissed import dismissed_keys
-from quodeq.shared._env import get_score_cache_path, score_cache_disabled
+from quodeq.services._score_cache_epoch import CACHE_WRITER_EPOCH as _CACHE_WRITER_EPOCH
 from quodeq.data.fs.suppression_rules import load_suppression_rules
 
-_logger = logging.getLogger(__name__)
-_BUSY_TIMEOUT_MS = 5000
-# Bumped when the cache *writer* semantics change in a way that could have
-# produced bad rows, to invalidate everything written by the prior writer.
-# "2": earlier writers persisted in-progress runs' partial scalar sets (e.g. 1
-# of 6 dims), which the content-hash version could never invalidate; the
-# write-guard now persists only completed runs, and this bump rebuilds the
-# stranded partial rows once.
-# "3": accumulated / project-summary payloads written before configured-dim
-# scoping carried stale dimensions (e.g. clean-architecture) that the project
-# no longer evaluates; the run-fingerprint could never invalidate them, so this
-# bump rebuilds them once against the latest run's configured-dimension set.
-# "4": earlier writers persisted in-progress runs' PARTIAL run_keys sets (the
-# per-run version path had no completeness gate), which load_run_keys froze
-# forever; the gate now persists only terminal runs, and this bump purges the
-# non-version-keyed run_keys table once so stranded partial snapshots rebuild.
-# "5": dismiss/delete rescoring switched basis from the legacy report-JSON
-# formula to the run's own evidence jsonl (services/evidence_rescore), so
-# scores cached by the prior writer differ for the SAME suppression state and
-# params; this bump rebuilds them on the evidence basis once.
-# "6": three scoring read paths built their run-dimension fetcher without the
-# staleness guards (in-progress bypass + eval-file-count validation), so a
-# request landing mid-run could freeze a partial dim list in the process LRU;
-# later writes built from it persisted half-rescored accumulated payloads and
-# partial scalar sets whose version hash can never self-invalidate. The guards
-# now live in the shared fetcher and the accumulated writer refuses
-# partial-coverage payloads; this bump rebuilds rows the prior writer may have
-# poisoned.
-_CACHE_WRITER_EPOCH = "6"
-
-# Shared-root isolation seam (Phase 2): when serving read endpoints from a
-# second (shared) clone, the score cache must not mix rows with the local
-# clone's cache. Unset (None) in every normal code path, so the default
-# behavior below is byte-identical to before this seam existed.
-_CACHE_PATH_OVERRIDE: ContextVar[str | None] = ContextVar(
-    "score_cache_path_override", default=None
+# ---------------------------------------------------------------------------
+# Decomposed submodules. Every moved name is re-exported here so external
+# callers and test patch targets keep working against this module.
+# ---------------------------------------------------------------------------
+from quodeq.services._score_cache_db import (  # noqa: F401 — facade re-export
+    open_score_cache,
+    score_cache_path_override,
 )
-
-
-@contextmanager
-def score_cache_path_override(path: str | Path) -> Iterator[None]:
-    """Route score-cache reads/writes to *path* instead of the default DB.
-
-    Active only for the duration of the ``with`` block (and any code it
-    calls, via contextvars' task-local propagation); always restored,
-    including when the block raises.
-    """
-    token = _CACHE_PATH_OVERRIDE.set(str(path))
-    try:
-        yield
-    finally:
-        _CACHE_PATH_OVERRIDE.reset(token)
-
-
-_SCHEMA = (
-    "CREATE TABLE IF NOT EXISTS run_scalars ("
-    " project TEXT NOT NULL, run_id TEXT NOT NULL, version TEXT NOT NULL,"
-    " dimension TEXT NOT NULL, overall_score TEXT, overall_grade TEXT,"
-    " updated_at TEXT NOT NULL DEFAULT (datetime('now')),"
-    " PRIMARY KEY (project, run_id, dimension, version));"
-    "CREATE INDEX IF NOT EXISTS idx_run_scalars_lookup ON run_scalars(project, version);"
-    "CREATE TABLE IF NOT EXISTS accumulated_cache ("
-    " project TEXT NOT NULL, version TEXT NOT NULL, payload TEXT NOT NULL,"
-    " updated_at TEXT NOT NULL DEFAULT (datetime('now')),"
-    " PRIMARY KEY (project, version));"
-    "CREATE TABLE IF NOT EXISTS project_summary_cache ("
-    " project TEXT PRIMARY KEY, version TEXT NOT NULL, payload TEXT NOT NULL);"
-    "CREATE TABLE IF NOT EXISTS run_keys ("
-    " project TEXT NOT NULL, run_id TEXT NOT NULL,"
-    " dismiss_keys TEXT NOT NULL, class_keys TEXT NOT NULL,"
-    " PRIMARY KEY (project, run_id));"
-    "CREATE TABLE IF NOT EXISTS cache_meta (key TEXT PRIMARY KEY, value TEXT NOT NULL);"
+from quodeq.services._score_cache_store import (  # noqa: F401 — facade re-export
+    load_run_keys,
+    read_cached_accumulated,
+    read_cached_project_summary,
+    read_cached_rows,
+    store_run_keys,
+    write_cached_accumulated,
+    write_cached_project_summary,
+    write_cached_rows,
 )
-
-
-def _purge_run_keys_on_epoch_change(conn: sqlite3.Connection) -> None:
-    """One-time purge of the non-version-keyed run_keys table on epoch bump.
-
-    ``run_scalars`` / ``accumulated_cache`` / ``project_summary_cache`` embed the
-    epoch in their version hash and self-invalidate, but ``run_keys`` rows are not
-    version-keyed, so a partial snapshot persisted by a prior writer would survive
-    a bump and stay frozen. Clearing the table once forces a fresh read.
-    """
-    try:
-        row = conn.execute(
-            "SELECT value FROM cache_meta WHERE key='writer_epoch'"
-        ).fetchone()
-        if row is not None and row[0] == _CACHE_WRITER_EPOCH:
-            return
-        conn.execute("DELETE FROM run_keys")
-        conn.execute(
-            "INSERT OR REPLACE INTO cache_meta (key, value) VALUES ('writer_epoch', ?)",
-            (_CACHE_WRITER_EPOCH,),
-        )
-        conn.commit()
-    except sqlite3.Error:
-        _logger.warning("run_keys epoch purge failed", exc_info=True)
-
-
-def _init(path: Path) -> sqlite3.Connection:
-    conn = sqlite3.connect(path)
-    try:
-        conn.execute("PRAGMA journal_mode = WAL")
-        conn.execute(f"PRAGMA busy_timeout = {_BUSY_TIMEOUT_MS}")
-        conn.executescript(_SCHEMA)
-        conn.commit()
-        _purge_run_keys_on_epoch_change(conn)
-    except sqlite3.DatabaseError:
-        # Close before re-raising so the caller's rebuild path can unlink the
-        # file with no open handle (Windows raises PermissionError otherwise).
-        conn.close()
-        raise
-    return conn
-
-
-# Serializes _init (and the corrupt-db rebuild). Concurrent first-opens of a
-# fresh DB race the schema DDL; the loser's lock error is indistinguishable
-# from corruption here, so the rebuild path would unlink the file out from
-# under the winner's live WAL connection (observed as a SIGBUS on the mmap'd
-# -shm). Only open/rebuild is serialized — yielded connections stay concurrent.
-_OPEN_LOCK = threading.Lock()
-
-
-@contextmanager
-def open_score_cache() -> Iterator[sqlite3.Connection]:
-    """Open the score cache DB (WAL). Rebuilds from scratch if corrupt/older-schema."""
-    override = _CACHE_PATH_OVERRIDE.get()
-    path = Path(override) if override else Path(get_score_cache_path())
-    path.parent.mkdir(parents=True, exist_ok=True)
-    with _OPEN_LOCK:
-        try:
-            conn = _init(path)
-        except sqlite3.DatabaseError:
-            _logger.warning("score cache at %s unreadable; rebuilding", path)
-            path.unlink(missing_ok=True)
-            conn = _init(path)
-    try:
-        yield conn
-    finally:
-        conn.close()
-
-
-def read_cached_rows(
-    conn: sqlite3.Connection, project: str, run_id: str, version: str,
-) -> list[DimensionResult] | None:
-    """Return cached scalar dims for (project, run_id, version), or None on miss/error."""
-    try:
-        rows = conn.execute(
-            "SELECT dimension, overall_score, overall_grade FROM run_scalars "
-            "WHERE project=? AND run_id=? AND version=? ORDER BY dimension",
-            (project, run_id, version),
-        ).fetchall()
-    except sqlite3.Error:
-        return None
-    if not rows:
-        return None
-    return [DimensionResult(dimension=r[0], overall_score=r[1], overall_grade=r[2]) for r in rows]
-
-
-def write_cached_rows(
-    conn: sqlite3.Connection, project: str, run_id: str, version: str,
-    dims: list[DimensionResult],
-) -> None:
-    """Replace all cached rows for (project, run_id) with *dims* at *version*.
-
-    Best-effort: logs and returns on any SQLite error (the caller still has the
-    computed result).
-    """
-    try:
-        conn.execute("DELETE FROM run_scalars WHERE project=? AND run_id=?", (project, run_id))
-        conn.executemany(
-            "INSERT OR REPLACE INTO run_scalars "
-            "(project, run_id, version, dimension, overall_score, overall_grade) "
-            "VALUES (?, ?, ?, ?, ?, ?)",
-            [(project, run_id, version, d.dimension, d.overall_score, d.overall_grade)
-             for d in dims if d.dimension],
-        )
-        conn.commit()
-    except sqlite3.Error:
-        _logger.warning("score cache write failed for %s/%s", project, run_id, exc_info=True)
-
-
-def read_cached_accumulated(
-    conn: sqlite3.Connection, project: str, version: str,
-) -> dict | None:
-    """Return the cached accumulated payload for (project, version), or None."""
-    try:
-        row = conn.execute(
-            "SELECT payload FROM accumulated_cache WHERE project=? AND version=?",
-            (project, version),
-        ).fetchone()
-    except sqlite3.Error:
-        return None
-    if row is None:
-        return None
-    try:
-        return json.loads(row[0])
-    except (ValueError, TypeError):
-        return None
-
-
-def write_cached_accumulated(
-    conn: sqlite3.Connection, project: str, version: str, payload: dict,
-) -> None:
-    """Replace the cached accumulated payload for *project* at *version*.
-
-    Single-slot per project: the DELETE clears any prior version first, so the
-    table holds at most one accumulated payload per project. The default
-    dashboard uses ``as_of=None`` (one version), so this is stable; rapidly
-    alternating distinct ``as_of`` historical views would each miss + overwrite.
-
-    Best-effort: logs and returns on any SQLite/serialization error.
-    """
-    try:
-        blob = json.dumps(payload)
-    except (TypeError, ValueError):
-        _logger.warning("accumulated payload for %s not serializable; skipping cache", project)
-        return
-    try:
-        conn.execute("DELETE FROM accumulated_cache WHERE project=?", (project,))
-        conn.execute(
-            "INSERT OR REPLACE INTO accumulated_cache (project, version, payload) VALUES (?, ?, ?)",
-            (project, version, blob),
-        )
-        conn.commit()
-    except sqlite3.Error:
-        _logger.warning("accumulated cache write failed for %s", project, exc_info=True)
+from quodeq.services._score_cache_fetch import (  # noqa: F401 — facade re-export
+    cached_accumulated,
+    cached_project_summary,
+    make_cache_backed_fetcher,
+)
+from quodeq.shared._env import get_score_cache_path  # noqa: F401 — facade re-export
 
 
 def _params_fingerprint(params: ScoringParams) -> str:
@@ -307,45 +103,6 @@ def run_scoped_version(
         "params": _params_fingerprint(params),
     }, sort_keys=True)
     return hashlib.sha256(payload.encode("utf-8")).hexdigest()
-
-
-def store_run_keys(
-    conn: sqlite3.Connection, project: str, run_id: str,
-    dismiss_keys: set[tuple], class_keys: set[tuple],
-) -> None:
-    """Persist a run's key sets (best-effort)."""
-    try:
-        conn.execute(
-            "INSERT OR REPLACE INTO run_keys (project, run_id, dismiss_keys, class_keys) "
-            "VALUES (?, ?, ?, ?)",
-            (project, run_id,
-             json.dumps(sorted(list(k) for k in dismiss_keys)),
-             json.dumps(sorted(list(k) for k in class_keys))),
-        )
-        conn.commit()
-    except sqlite3.Error:
-        _logger.warning("run_keys write failed for %s/%s", project, run_id, exc_info=True)
-
-
-def load_run_keys(
-    conn: sqlite3.Connection, project: str,
-) -> dict[str, tuple[set[tuple], set[tuple]]]:
-    """Return ``{run_id: (dismiss_keys, class_keys)}`` for *project* (empty on error)."""
-    out: dict[str, tuple[set[tuple], set[tuple]]] = {}
-    try:
-        rows = conn.execute(
-            "SELECT run_id, dismiss_keys, class_keys FROM run_keys WHERE project=?",
-            (project,),
-        ).fetchall()
-    except sqlite3.Error:
-        return {}
-    for run_id, dj, cj in rows:
-        try:
-            out[run_id] = ({tuple(k) for k in json.loads(dj)},
-                           {tuple(k) for k in json.loads(cj)})
-        except (ValueError, TypeError):
-            continue
-    return out
 
 
 def accumulated_cache_version(
@@ -437,201 +194,3 @@ def per_run_versions(
         out.append(
             (rid, status, run_scoped_version(params, keys[0], keys[1], dismissed, deleted)))
     return out
-
-
-# In-flight computes by (kind, project, version). Concurrent misses on the
-# same key must share ONE compute: these computes walk a project's full run
-# history and can take minutes right after an upgrade invalidates every
-# cached row, and the client re-requests while the first compute is still
-# running. The registry entry is dropped once no thread holds the key's lock;
-# a waiter that raced the cleanup at worst recomputes (idempotent write).
-_INFLIGHT_GUARD = threading.Lock()
-_INFLIGHT: dict[tuple[str, str, str], threading.Lock] = {}
-
-
-@contextmanager
-def _single_flight(kind: str, project: str, version: str) -> Iterator[None]:
-    key = (kind, project, version)
-    with _INFLIGHT_GUARD:
-        lock = _INFLIGHT.setdefault(key, threading.Lock())
-    lock.acquire()
-    try:
-        yield
-    finally:
-        lock.release()
-        with _INFLIGHT_GUARD:
-            if not lock.locked() and _INFLIGHT.get(key) is lock:
-                del _INFLIGHT[key]
-
-
-def cached_accumulated(
-    project: str, version: str, compute: Callable[[], dict],
-    cacheable: Callable[[dict], bool] | None = None,
-) -> dict:
-    """Read-through cache for the accumulated payload.
-
-    Hit -> return the deserialized cached payload. Miss (or kill switch / cache
-    error) -> call *compute*, cache the result best-effort, return it.
-    Concurrent misses on one (project, version) share a single compute.
-
-    *cacheable*, when given, is called with the computed result before it is
-    persisted; returning False serves the result without caching it. This lets
-    the caller withhold payloads it knows are incomplete (e.g. a rescore that
-    covered only part of the dimensions), which would otherwise freeze under a
-    version hash that cannot self-invalidate.
-    """
-    if score_cache_disabled():
-        return compute()
-    try:
-        with open_score_cache() as conn:
-            cached = read_cached_accumulated(conn, project, version)
-        if cached is not None:
-            return cached
-    except sqlite3.Error:
-        return compute()
-    with _single_flight("accumulated", project, version):
-        # Re-check: a caller we waited on may have computed and cached it.
-        try:
-            with open_score_cache() as conn:
-                cached = read_cached_accumulated(conn, project, version)
-            if cached is not None:
-                return cached
-        except sqlite3.Error:
-            pass
-        result = compute()
-        if cacheable is not None and not cacheable(result):
-            return result
-        try:
-            with open_score_cache() as conn:
-                write_cached_accumulated(conn, project, version, result)
-        except sqlite3.Error:
-            pass
-        return result
-
-
-def read_cached_project_summary(
-    conn: sqlite3.Connection, project: str, version: str,
-) -> dict | None:
-    """Return the cached project-card summary for (project, version), or None."""
-    try:
-        row = conn.execute(
-            "SELECT payload FROM project_summary_cache WHERE project=? AND version=?",
-            (project, version),
-        ).fetchone()
-    except sqlite3.Error:
-        return None
-    if row is None:
-        return None
-    try:
-        return json.loads(row[0])
-    except (ValueError, TypeError):
-        return None
-
-
-def write_cached_project_summary(
-    conn: sqlite3.Connection, project: str, version: str, payload: dict,
-) -> None:
-    """Single-slot-per-project write for the project-card summary."""
-    try:
-        blob = json.dumps(payload)
-    except (TypeError, ValueError):
-        return
-    try:
-        conn.execute("DELETE FROM project_summary_cache WHERE project=?", (project,))
-        conn.execute(
-            "INSERT OR REPLACE INTO project_summary_cache (project, version, payload) VALUES (?, ?, ?)",
-            (project, version, blob),
-        )
-        conn.commit()
-    except sqlite3.Error:
-        _logger.warning("project summary cache write failed for %s", project, exc_info=True)
-
-
-def cached_project_summary(
-    project: str, version: str, compute: Callable[[], dict],
-) -> dict:
-    """Read-through cache for the project-card summary (mirrors cached_accumulated)."""
-    if score_cache_disabled():
-        return compute()
-    try:
-        with open_score_cache() as conn:
-            hit = read_cached_project_summary(conn, project, version)
-        if hit is not None:
-            return hit
-    except sqlite3.Error:
-        return compute()
-    with _single_flight("summary", project, version):
-        # Re-check: a caller we waited on may have computed and cached it.
-        try:
-            with open_score_cache() as conn:
-                hit = read_cached_project_summary(conn, project, version)
-            if hit is not None:
-                return hit
-        except sqlite3.Error:
-            pass
-        result = compute()
-        try:
-            with open_score_cache() as conn:
-                write_cached_project_summary(conn, project, version, result)
-        except sqlite3.Error:
-            pass
-        return result
-
-
-def make_cache_backed_fetcher(
-    project: str, version_for: Callable[[str], str],
-    base_fetcher: Callable[[str], list[DimensionResult]],
-    is_cacheable: Callable[[str], bool] | None = None,
-) -> Callable[[str], list[DimensionResult]]:
-    """Wrap *base_fetcher* with the read-through cache, versioned PER RUN.
-
-    *version_for(run_id)* returns that run's scoped version (params + the
-    suppressions touching it). Bulk-loads every cached row for *project* keyed by
-    (run_id, version); a hit requires the row's version to equal the run's
-    current version, so a dismiss/delete only misses the runs it touches. Misses
-    compute via *base_fetcher*, cache scalars at the run's version (only when
-    ``is_cacheable``), and return them. Kill switch -> *base_fetcher* unchanged.
-
-    ``is_cacheable`` gates *persistence* per run: only terminal (complete) runs
-    are safe. An in-progress run's scalar set grows as dimensions finish, and the
-    version hash can't see that, so persisting its partial set would strand a
-    stale row -- so opening History mid-scan would leave the trend showing one
-    dimension forever while run-detail shows all six. Non-cacheable runs
-    compute-through and are served for the current build but never written to
-    disk. Defaults to "always cacheable" for backward compatibility.
-    """
-    if score_cache_disabled():
-        return base_fetcher
-
-    by_run_version: dict[tuple[str, str], list[DimensionResult]] = {}
-    try:
-        with open_score_cache() as conn:
-            for rid, ver, dim, score, grade in conn.execute(
-                "SELECT run_id, version, dimension, overall_score, overall_grade "
-                "FROM run_scalars WHERE project=? ORDER BY run_id, dimension",
-                (project,),
-            ):
-                by_run_version.setdefault((rid, ver), []).append(
-                    DimensionResult(dimension=dim, overall_score=score, overall_grade=grade))
-    except sqlite3.Error:
-        by_run_version = {}
-
-    def fetch(run_id: str) -> list[DimensionResult]:
-        version = version_for(run_id)
-        hit = by_run_version.get((run_id, version))
-        if hit is not None:
-            return hit
-        dims = base_fetcher(run_id)
-        scalars = [DimensionResult(dimension=d.dimension, overall_score=d.overall_score,
-                                   overall_grade=d.overall_grade)
-                   for d in dims if d.dimension]
-        by_run_version[(run_id, version)] = scalars
-        if is_cacheable is None or is_cacheable(run_id):
-            try:
-                with open_score_cache() as conn:
-                    write_cached_rows(conn, project, run_id, version, scalars)
-            except sqlite3.Error:
-                pass
-        return scalars
-
-    return fetch
